Fix JSON Roundtrip assert for .NET 11 exact double-to-decimal conversion

+++ test/EFCore.MySql.FunctionalTests/Query/JsonNewtonsoftDomQueryTest.cs
@@ -49,7 +49,11 @@ namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.Query
 
                  var order2 = customer["Orders"][1];
 
-                 Assert.Equal(23.1m, order2["Price"].Value<decimal>());
+                 // Newtonsoft parses JSON numbers as doubles. Since .NET 11, converting a double
+                 // to decimal keeps the exact binary value instead of rounding to 15 significant
+                 // digits, so 23.1 now converts to 23.100000000000001421085471520. 99.5 above is
+                 // unaffected because it is exactly representable in binary floating point.
+                 Assert.Equal(23.1, order2["Price"].Value<double>());
                  Assert.Equal("Some address 2", order2["ShippingAddress"].Value<string>());
                  Assert.Equal(new DateTime(2019, 10, 10), order2["ShippingDate"].Value<DateTime>());
              }
